requirements.txt
requests
pandas
matplotlib
snowflake-connector-python
python-dotenv
